Style : Seperate roomlist to main
Seperate roomlist to create roomList component.
