+++ main/src/component/main/select.js
@@ -17,7 +17,7 @@ const Select = (props)=>{
     const [genre, setGenre] = useState('');
 
     const maxNum = async (genre)=>{
-        const recipeUrl = "https://3001-orange-vicuna-9uo5wxk0.ws-us09.gitpod.io/room/genres";
+        const recipeUrl = "https://3001-orange-vicuna-9uo5wxk0.ws-us08.gitpod.io/room/genres";
         const requestMetadata = {
             method: 'POST',
             headers: {
@@ -72,7 +72,7 @@ const Select = (props)=>{
     
 	// when Add room=> dispatch room
     const sendServer = (room)=>{
-        const recipeUrl = "https://3001-orange-vicuna-9uo5wxk0.ws-us09.gitpod.io/room/add";
+        const recipeUrl = "https://3001-orange-vicuna-9uo5wxk0.ws-us08.gitpod.io/room/add";
         const requestMetadata = {
             method: 'POST',
             headers: {
